requirements.txt
pyyaml
scp
paramiko
libvirt-python
click